% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{add_x_axis}
\alias{add_x_axis}
\title{Show X-Axis}
\usage{
add_x_axis(title = TRUE, text = TRUE, ticks = TRUE, line = TRUE)
}
\arguments{
\item{title}{Logical; include axis title?}

\item{text}{Logical; include axis text/labels?}

\item{ticks}{Logical; include axis ticks?}

\item{line}{Logical; include axis line?}
}
\value{
A ggplot2 theme element.
}
\description{
Ensure x-axis elements are visible in a ggplot2 plot.
}
