% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_column.R
\name{as_ggplot}
\alias{as_ggplot}
\title{Convert PlotColumn to ggplot2 Objects}
\usage{
as_ggplot(x, ...)
}
\arguments{
\item{x}{A \code{\link[=plot_column]{plot_column()}} or other object.}

\item{...}{Additional arguments (unused).}
}
\value{
A list of ggplot2 objects.
}
\description{
Generic function to extract the underlying ggplot2 objects.
}
