% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_column.R
\name{as_html}
\alias{as_html}
\title{Convert PlotColumn to HTML}
\usage{
as_html(x, ...)
}
\arguments{
\item{x}{A \code{\link[=plot_column]{plot_column()}} or other object.}

\item{...}{Additional arguments (unused).}
}
\value{
A list of \link[htmltools:builder]{htmltools::tags} objects.
}
\description{
Generic function to extract HTML representations from a PlotColumn.
}
