% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_column.R
\name{as_R6}
\alias{as_R6}
\title{Extract R6 Objects from PlotColumn}
\usage{
as_R6(x, ...)
}
\arguments{
\item{x}{A \code{\link[=plot_column]{plot_column()}} or other object.}

\item{...}{Additional arguments (unused).}
}
\value{
A list of \link{SvgPlot} R6 objects.
}
\description{
Generic function to extract the underlying R6 \link{SvgPlot} objects.
}
