% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_text_color}
\alias{get_text_color}
\title{Get Contrasting Text Color}
\usage{
get_text_color(background_color)
}
\arguments{
\item{background_color}{Character vector of background color(s).}
}
\value{
Character vector of \code{"black"} or \code{"white"}.
}
\description{
Returns \code{"white"} for dark backgrounds, \code{"black"} for light backgrounds.
}
\examples{
plotweave::get_text_color("#007DBA") # "white"
plotweave::get_text_color("#FFFFFF") # "black"
}
