% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getfun}
\alias{getfun}
\title{Get function from character string}
\usage{
getfun(x)
}
\arguments{
\item{x}{Character string or function. If already a function, returned as-is.}
}
\value{
Function object.
}
\description{
Resolve a function from a character string, supporting namespace-qualified
names like \code{"pkg::fun"}.
}
\keyword{internal}
