% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{hide_legend}
\alias{hide_legend}
\title{Hide Plot Legend}
\usage{
hide_legend()
}
\value{
A ggplot2 theme element.
}
\description{
Remove the legend from a ggplot2 plot.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, color = factor(cyl))) +
  geom_point() +
  plotweave::hide_legend()
}
