% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{hide_title}
\alias{hide_title}
\title{Hide Plot Title}
\usage{
hide_title()
}
\value{
A ggplot2 theme element.
}
\description{
Remove the title and subtitle from a ggplot2 plot.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  ggtitle("My Plot") +
  plotweave::hide_title()
}
