% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{hide_x_axis}
\alias{hide_x_axis}
\title{Hide X-Axis}
\usage{
hide_x_axis()
}
\value{
A ggplot2 theme element.
}
\description{
Remove all x-axis elements (text, title, ticks, line) from a ggplot2 plot.
}
