% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{hide_y_axis}
\alias{hide_y_axis}
\title{Hide Y-Axis}
\usage{
hide_y_axis()
}
\value{
A ggplot2 theme element.
}
\description{
Remove all y-axis elements (text, title, ticks, line) from a ggplot2 plot.
}
