% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_column.R
\name{is_plot_column}
\alias{is_plot_column}
\title{Check if an Object is a PlotColumn}
\usage{
is_plot_column(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical.
}
\description{
Check if an Object is a PlotColumn
}
