% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_column.R
\name{plot_column}
\alias{plot_column}
\title{Create a PlotColumn Vector}
\usage{
plot_column(...)
}
\arguments{
\item{...}{\link{SvgPlot} objects to include in the column.}
}
\value{
A \code{PlotColumn} vctrs vector.
}
\description{
A vctrs-based column type for embedding \link{SvgPlot} objects in data frames
and tables. Each element is an \link{SvgPlot} R6 instance.
}
