% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_column.R
\name{plot_height}
\alias{plot_height}
\title{Get Maximum Plot Height}
\usage{
plot_height(x, ...)
}
\arguments{
\item{x}{A \code{\link[=plot_column]{plot_column()}} vector.}

\item{...}{Additional arguments (unused).}
}
\value{
Numeric maximum height.
}
\description{
Returns the maximum SVG viewBox height across all plots in a PlotColumn.
}
