% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_column.R
\name{plot_width}
\alias{plot_width}
\title{Get Maximum Plot Width}
\usage{
plot_width(x, ...)
}
\arguments{
\item{x}{A \code{\link[=plot_column]{plot_column()}} vector.}

\item{...}{Additional arguments (unused).}
}
\value{
Numeric maximum width.
}
\description{
Returns the maximum SVG viewBox width across all plots in a PlotColumn.
}
