% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plotweave_colors}
\alias{plotweave_colors}
\title{Plotweave Color Palette}
\usage{
plotweave_colors()
}
\value{
Named character vector of hex color codes.
}
\description{
Returns the full named vector of plotweave colors. These serve as default
colors for inline plots and can be overridden via \code{\link[=pw_set_config]{pw_set_config()}}.
}
\examples{
plotweave::plotweave_colors()
plotweave::plotweave_colors()["blue"]
}
