% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plotweave_palette_discrete}
\alias{plotweave_palette_discrete}
\title{Plotweave Discrete Color Palette}
\usage{
plotweave_palette_discrete()
}
\value{
A function with signature \verb{function(n)} returning \code{n} hex colors.
}
\description{
Returns a function that, given \code{n}, produces a character vector of \code{n}
colors drawn from the plotweave palette.
}
\examples{
pal <- plotweave::plotweave_palette_discrete()
pal(3)
}
