% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pop}
\alias{pop}
\title{Pop elements from a vector}
\usage{
pop(x, to_remove)
}
\arguments{
\item{x}{Character vector to filter.}

\item{to_remove}{Elements to remove from \code{x}.}
}
\value{
Character vector with elements removed.
}
\description{
Remove specified elements from a character vector.
}
\keyword{internal}
