% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{pw_config}
\alias{pw_config}
\title{Get Plotweave Configuration}
\usage{
pw_config()
}
\value{
A list with fields: \code{font_family}, \code{font_url}, \code{colors},
\code{missing_color}, \code{default_width}, \code{default_height}, \code{default_scaling},
\code{download_button}.
}
\description{
Returns the current global configuration as a named list.
}
\examples{
plotweave::pw_config()
}
