% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{pw_reset_config}
\alias{pw_reset_config}
\title{Reset Plotweave Configuration}
\usage{
pw_reset_config()
}
\value{
Invisible list of the default configuration.
}
\description{
Restore all configuration values to package defaults.
}
\examples{
plotweave::pw_reset_config()
}
